テトリミノ作成試し

+++ Tetris/Def.h
@@ -2,8 +2,8 @@
 
 // define
 static constexpr float PI = 3.14159265358f;
-static Vec2f WINDOW_SIZE = { 640.f, 480.f };
-static float ASPECT_RATIO = WINDOW_SIZE.x / WINDOW_SIZE.y;
+static Vec2i WINDOW_SIZE = { 640, 480 };
+static float ASPECT_RATIO = static_cast<float>(WINDOW_SIZE.x) / WINDOW_SIZE.y;
 //static constexpr Vec2i FIELD_SIZE = { 10, 20 };
 
 // undef
